made important changes in backend to update & delete tasks by their respective sequence.

+++ server/controllers/tasks/deleteTasksController.js
@@ -1,7 +1,8 @@
 const UsersModel = require("../../mongodb/models/usersModel");
 const WorkspaceModel = require('../../mongodb/models/workspacesModel');
 const TasksModel = require("../../mongodb/models/tasksModel");
-const { isValidObjectIdString } = require("../../libs/helperFunctions");
+const SectionsModel = require("../../mongodb/models/sectionsModel");
+const { isValidObjectIdString, insertValueAtIndex } = require("../../libs/helperFunctions");
 
 const deleteTasksController = async (req, res) => {
     let status = 200;
@@ -11,8 +12,8 @@ const deleteTasksController = async (req, res) => {
     }
     try {
 
-        const { task_id, workspace_id, user_id } = req.body;
-        if (task_id && workspace_id && user_id) {
+        const { task_id, section_id, workspace_id, user_id } = req.body;
+        if (task_id && section_id && workspace_id && user_id) {
             const workspaceIDCheck = isValidObjectIdString(workspace_id);
             const verTok = req.user.user_id;
             // Check user already exist.
@@ -21,19 +22,60 @@ const deleteTasksController = async (req, res) => {
                 if (workspaceIDCheck) {
                     const workspaceAlreadyExist = await WorkspaceModel.findOne({ _id: workspace_id });
                     if (workspaceAlreadyExist !== null) {
-                        const taskAlreadyExist = await TasksModel.findOne({ _id: task_id });
-                        if (taskAlreadyExist !== null) {
-                            await TasksModel.findByIdAndDelete({ _id: task_id });
-                            status = 200;
-                            response = {
-                                success: true,
-                                message: "Task deleted successfully."
+                        const sectionIDCheck = isValidObjectIdString(section_id);
+                        if (sectionIDCheck) {
+                            const sectionAlreadyExist = await SectionsModel.findOne({ _id: section_id });
+                            if (sectionAlreadyExist !== null) {
+                                const taskIDCheck = isValidObjectIdString(task_id);
+                                if (taskIDCheck) {
+                                    const taskAlreadyExist = await TasksModel.findOne({ _id: task_id });
+                                    if (taskAlreadyExist !== null) {
+                                        if (taskAlreadyExist.section_id == section_id) {
+                                            await TasksModel.findByIdAndDelete({ _id: task_id });
+                                            const filr = sectionAlreadyExist.task_sequence.filter((ids) => ids !== task_id);
+                                            await SectionsModel.findByIdAndUpdate({ _id: section_id }, { task_sequence: filr })
+                                            status = 200;
+                                            response = {
+                                                success: true,
+                                                message: "Task deleted successfully."
+                                            }
+                                        } else {
+                                            const newSection = await SectionsModel.findOne({ _id: section_id });
+                                            const valueAlreadyExist = newSection.task_sequence.includes(task_id);
+                                            if (!valueAlreadyExist) {
+                                                status = 200;
+                                                response = {
+                                                    success: false,
+                                                    message: "Please save the changes before deleting task."
+                                                }
+                                            }
+                                        }
+                                    } else {
+                                        status = 200;
+                                        response = {
+                                            success: false,
+                                            message: "Task not found."
+                                        }
+                                    }
+                                } else {
+                                    status = 200;
+                                    response = {
+                                        success: false,
+                                        message: "Invalid task ID found."
+                                    }
+                                }
+                            } else {
+                                status = 200;
+                                response = {
+                                    success: false,
+                                    message: "Section not found."
+                                }
                             }
                         } else {
                             status = 200;
                             response = {
                                 success: false,
-                                message: "Task not found."
+                                message: "Invalid section ID found."
                             }
                         }
                     } else {
